+++ ui/app/unlock.js
@@ -72,7 +72,7 @@ UnlockScreen.prototype.render = function () {
           style: {
             margin: 10,
           },
-        }, 'Unlock'),
+        }, 'Log In'),
       ]),
 
       h('.flex-row.flex-center.flex-grow', [
@@ -105,7 +105,7 @@ UnlockScreen.prototype.render = function () {
           },
         }, 'Use classic interface'),
       ]),
-      
+
     ])
   )
 }